Added Plot_linescans
We've added a new script that is a modified version of Zac Swider's wave analysis script. This updated version reads kymographs and creates linescans across each column of pixels, allowing us to measure the characteristics of waves. This feature is particularly useful for movies captured at the edge of cells where there is only a line of wave activity. It offers an effective way to analyze such data.